+++ examples/next-image/pages/about.js
@@ -1,0 +1,14 @@
+import Head from 'next/head'
+import Link from 'next/link'
+
+const About = () => (
+  <>
+    <Head>
+      <title>next-pwa example</title>
+    </Head>
+    <h1>About</h1>
+    <Link href='/'>Home</Link>
+  </>
+)
+
+export default About