+++ index.js
@@ -217,6 +217,7 @@ module.exports = (nextConfig = {}) => ({
           ...manifestTransforms,
           async (manifestEntries, compilation) => {
             const manifest = manifestEntries.map(m => {
+              m.url = m.url.replace('/_next//static/image', '/_next/static/image')
               m.url = m.url.replace(/\[/g, '%5B').replace(/\]/g, '%5D')
               m.revision = buildId
               return m